+++ script/index.js
@@ -1,5 +1,5 @@
 export { default as build } from './build.js'
 export { default as lint } from './lint.js'
 export { default as test } from './test.js'
-export { default as publish } from './publish.js'
+export { default as release } from './release.js'
 export { default as update } from './update.js'