+++ script/update.js
@@ -1,5 +1,9 @@
+import { existsSync, unlinkSync } from 'fs'
+import { join } from 'path'
 import ncu from 'npm-check-updates'
-import { log } from '../utility/log.js'
+import { execSync } from 'child_process'
+import rimraf from 'rimraf'
+import log from 'logua'
 
 export default async () => {
   log('checking for updates..')
@@ -12,8 +16,6 @@ export default async () => {
     return log('everything already up-to-date')
   }
 
-  log('upgrade done')
-
   Object.keys(upgrades).forEach((key) => {
     const version = upgrades[key]
 
@@ -21,4 +23,21 @@ export default async () => {
   })
 
   console.log('')
+
+  log('dependencies upgraded in package.json')
+
+  log('reinstalling dependencies after upgrade...')
+
+  // Cleanup before install.
+  rimraf.sync('node_modules')
+
+  const packageLockFilePath = join(process.cwd(), 'package-lock.json')
+
+  if (existsSync(packageLockFilePath)) {
+    unlinkSync(packageLockFilePath)
+  }
+
+  execSync('npm install', { stdio: 'inherit' })
+
+  log('new dependencies installed')
 }