+++ CHANGELOG.md
@@ -2,6 +2,48 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.0.7](https://github.com/tobua/padua/compare/v0.0.6...v0.0.7) (2020-09-01)
+
+
+### Features
+
+* **general:** various improvements ([0ee6479](https://github.com/tobua/padua/commit/0ee6479593c21f28552f87f68fb57bec6de6c30c))
+
+### [0.0.6](https://github.com/tobua/padua/compare/v0.0.5...v0.0.6) (2020-08-28)
+
+### [0.0.5](https://github.com/tobua/padua/compare/v0.0.4...v0.0.5) (2020-08-24)
+
+
+### Features
+
+* **template:** add and improve node templates ([d191931](https://github.com/tobua/padua/commit/d1919314e4f40dc06e844acd92cef5b2ceb5fa97))
+
+### [0.0.4](https://github.com/tobua/padua/compare/v0.0.3...v0.0.4) (2020-08-21)
+
+
+### Features
+
+* **install:** sort package.json properties on install ([9c503e3](https://github.com/tobua/padua/commit/9c503e3dec4c7bcc7b2712d5d1a6ba218feaf54a))
+
+
+### Bug Fixes
+
+* **lint:** configure jest globals for test files ([b1f1b99](https://github.com/tobua/padua/commit/b1f1b993a1c4b95f334b864c8964569010d156ae))
+
+### [0.0.3](https://github.com/tobua/padua/compare/v0.0.2...v0.0.3) (2020-08-19)
+
+
+### Bug Fixes
+
+* **test:** run JS through babel-jest and add configuration ([a2ef7ca](https://github.com/tobua/padua/commit/a2ef7ca545d896c8da6e511a51fc3a096efa87e9))
+
+### [0.0.2](https://github.com/tobua/padua/compare/v0.0.1...v0.0.2) (2020-08-18)
+
+
+### Features
+
+* **template:** add repository field and default template ([56cfa0b](https://github.com/tobua/padua/commit/56cfa0bf8e49c7c28111f8a22ad2b0eee8793b9e))
+
 ### [0.0.1](https://github.com/tobua/padua/compare/v0.0.0...v0.0.1) (2020-08-05)